🔥 chore: move config file to infrastructure
