+++ src/config/index.ts
@@ -8,28 +8,17 @@ const Config = {
   DATABASE_URL: process.env.MONGODB_URI!,
   JWT_SECRET: process.env.JWT_SECRET_KEY!,
   SERVER_URL: process.env.ORIGIN_SERVER_URL!,
-  MAIN_SERVER_URL: process.env.MAIN_SERVER_URL!,
-  BUILD_SERVER_URL: process.env.BUILD_SERVER_URL!,
-  MAIN_SERVER_PORT: process.env.MAIN_SERVER_PORT!,
-  BUILD_SERVER_PORT: process.env.BUILD_SERVER_PORT!,
-
+  SERVER_PORT: process.env.SERVER_PORT!,
   CLIENT_URL: process.env.CLIENT_URL!,
+
+  // 변경중
   CLIENT_ID: process.env.GITHUB_CLIENT_ID!,
   CLIENT_SECRET: process.env.GITHUB_CLIENT_SECRET!,
   USER_CREDENTIAL_ID: process.env.GITHUB_USER_CREDENTIAL_ID!,
   USER_CREDENTIAL_TOKEN: process.env.GITHUB_USER_CREDENTIAL_TOKEN!,
   WEBHOOK_SECRET_KEY: process.env.GITHUB_WEBHOOK_SECRET_KEY!,
   WEBHOOK_PAYLOAD_URL: process.env.GITHUB_WEBHOOK_PAYLOAD_URL!,
-  INSTANCE_TYPE: process.env.AWS_INSTANCE_TYPE!,
-  AMI_ID: process.env.AWS_AMI_ID!,
-  KEY_PAIR_NAME: process.env.AWS_KEY_PAIR_NAME!,
-  IAM_INSTANCE_PROFILE: process.env.AWS_IAM_INSTANCE_PROFILE!,
-  INSTANCE_REGION: process.env.AWS_INSTANCE_REGION!,
-  HOSTED_ZONE_ID: process.env.AWS_HOSTED_ZONE_ID!,
-  ACCESS_KEY_ID: process.env.AWS_ACCESS_KEY_ID!,
-  SECRET_ACCESS_KEY: process.env.AWS_SECRET_ACCESS_KEY!,
-  NODEJS_FERMIUM: "14.21.0",
-  NODEJS_GALLIUM: "16.18.0",
+
   CLIENT_OPTIONS: {
     debug: true,
   },
